[Cache] Moving Cache classes to a new Manager namespace for better organization.
