+++ src/Valkyrja/Application/Env.php
@@ -27,8 +27,8 @@ use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Auth\Store\OrmStore;
 use Valkyrja\Broadcast\Contract\Broadcaster;
 use Valkyrja\Broadcast\PusherBroadcaster;
-use Valkyrja\Cache\Contract\Cache;
-use Valkyrja\Cache\RedisCache;
+use Valkyrja\Cache\Manager\Contract\Cache;
+use Valkyrja\Cache\Manager\RedisCache;
 use Valkyrja\Cli\Middleware\Contract\CommandDispatchedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\CommandMatchedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\CommandNotMatchedMiddleware;
