+++ src/Valkyrja/Cache/Manager/Contract/Cache.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cache\Contract;
+namespace Valkyrja\Cache\Manager\Contract;
 
 use Valkyrja\Cache\Tagger\Contract\Tagger;
