+++ src/Valkyrja/Cache/Manager/RedisCache.php
@@ -11,31 +11,29 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cache;
+namespace Valkyrja\Cache\Manager;
 
-use JsonException;
 use Override;
-use Valkyrja\Cache\Contract\Cache as Contract;
+use Predis\Client;
+use Valkyrja\Cache\Manager\Contract\Cache as Contract;
 use Valkyrja\Cache\Tagger\Contract\Tagger;
 use Valkyrja\Cache\Tagger\Tagger as TagClass;
-use Valkyrja\Log\Logger\Contract\Logger;
-use Valkyrja\Type\BuiltIn\Support\Arr;
 
 /**
- * Class LogCache.
+ * Class RedisCache.
  *
  * @author Melech Mizrachi
  */
-class LogCache implements Contract
+class RedisCache implements Contract
 {
     /**
-     * LogCache constructor.
+     * RedisCache constructor.
      *
-     * @param Logger $logger The logger service
-     * @param string $prefix [optional] The prefix
+     * @param Client $client The predis client
+     * @param string $prefix The prefix
      */
     public function __construct(
-        protected Logger $logger,
+        protected Client $client,
         protected string $prefix = ''
     ) {
     }
@@ -46,9 +44,7 @@ class LogCache implements Contract
     #[Override]
     public function has(string $key): bool
     {
-        $this->logger->info(self::class . " has: $key");
-
-        return true;
+        return (bool) $this->client->exists($this->getKey($key));
     }
 
     /**
@@ -57,24 +53,24 @@ class LogCache implements Contract
     #[Override]
     public function get(string $key): string|null
     {
-        $this->logger->info(self::class . " get: $key");
-
-        return '';
+        return $this->client->get($this->getKey($key));
     }
 
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @psalm-suppress MixedReturnTypeCoercion
      */
     #[Override]
     public function many(string ...$keys): array
     {
-        $keysString = Arr::toString($keys);
+        $prefixedKeys = [];
 
-        $this->logger->info(self::class . " many: $keysString");
+        foreach ($keys as $key) {
+            $prefixedKeys[] = $this->getKey($key);
+        }
 
-        return [];
+        return $this->client->mget($prefixedKeys);
     }
 
     /**
@@ -83,20 +79,24 @@ class LogCache implements Contract
     #[Override]
     public function put(string $key, string $value, int $minutes): void
     {
-        $this->logger->info(self::class . " put: $key, value $value, minutes $minutes");
+        $this->client->setex($this->getKey($key), $minutes * 60, $value);
     }
 
     /**
      * @inheritDoc
-     *
-     * @throws JsonException
      */
     #[Override]
     public function putMany(array $values, int $minutes): void
     {
-        $valuesString = Arr::toString($values);
+        $seconds = $minutes * 60;
 
-        $this->logger->info(self::class . " putMany: $valuesString, minutes $minutes");
+        $this->client->transaction(
+            function (Client $client) use ($values, $seconds): void {
+                foreach ($values as $key => $value) {
+                    $client->setex($this->getKey($key), $seconds, $value);
+                }
+            }
+        );
     }
 
     /**
@@ -105,9 +105,7 @@ class LogCache implements Contract
     #[Override]
     public function increment(string $key, int $value = 1): int
     {
-        $this->logger->info(self::class . " increment: $key, value $value");
-
-        return $value;
+        return $this->client->incrby($this->getKey($key), $value);
     }
 
     /**
@@ -116,9 +114,7 @@ class LogCache implements Contract
     #[Override]
     public function decrement(string $key, int $value = 1): int
     {
-        $this->logger->info(self::class . " decrement: $key, value $value");
-
-        return $value;
+        return $this->client->decrby($this->getKey($key), $value);
     }
 
     /**
@@ -127,7 +123,7 @@ class LogCache implements Contract
     #[Override]
     public function forever(string $key, $value): void
     {
-        $this->logger->info(self::class . " forever: $key, value $value");
+        $this->client->set($this->getKey($key), $value);
     }
 
     /**
@@ -136,9 +132,7 @@ class LogCache implements Contract
     #[Override]
     public function forget(string $key): bool
     {
-        $this->logger->info(self::class . " forget: $key");
-
-        return true;
+        return (bool) $this->client->del([$this->getKey($key)]);
     }
 
     /**
@@ -147,9 +141,7 @@ class LogCache implements Contract
     #[Override]
     public function flush(): bool
     {
-        $this->logger->info(self::class . ' flush');
-
-        return true;
+        return (bool) $this->client->flushdb();
     }
 
     /**
