+++ src/Valkyrja/Cache/Provider/ServiceProvider.php
@@ -16,10 +16,10 @@ namespace Valkyrja\Cache\Provider;
 use Override;
 use Predis\Client;
 use Valkyrja\Application\Env;
-use Valkyrja\Cache\Contract\Cache;
-use Valkyrja\Cache\LogCache;
-use Valkyrja\Cache\NullCache;
-use Valkyrja\Cache\RedisCache;
+use Valkyrja\Cache\Manager\Contract\Cache;
+use Valkyrja\Cache\Manager\LogCache;
+use Valkyrja\Cache\Manager\NullCache;
+use Valkyrja\Cache\Manager\RedisCache;
 use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
 use Valkyrja\Log\Logger\Contract\Logger;
